integrating kernel

+++ tpu_inference/layers/vllm/linear_common.py
@@ -23,7 +23,11 @@ def sharded_quantized_matmul(x: jax.Array, w_q: jax.Array, w_s: jax.Array,
     x = jax.lax.with_sharding_constraint(x, NamedSharding(mesh, x_sharding))
 
     def wrapper(x, w_q, w_s):
-        output = quantized_matmul_kernel(x, w_q, w_s, sc_size=128)
+        output = quantized_matmul_kernel(x,
+                                         w_q,
+                                         w_s,
+                                         x_q_dtype=w_q.dtype,
+                                         sc_size=128)
         if in_axis:
             output = jax.lax.psum(output, axis_name=in_axis)
         return output
